split intro widgets
