+++ focusfeed/lib/main.dart
@@ -6,7 +6,7 @@ import 'features/auth/screens/app_entry_screen.dart';
 import 'features/auth/screens/create_account_screen.dart';
 import 'features/profile/screens/profile_setup_screen.dart';
 import 'features/nav/main_nav_screen.dart';
-import 'features/intro/intro_screen.dart';
+import 'features/intro/screens/intro_screen.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
